requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.9.1
SQLAlchemy==2.0.34
aiosqlite==0.20.0
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
ulid-py==1.1.0
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
odfpy==1.4.1
